Move polymorphism tests into package
