+++ packages/tanstack-react-query/test/polymorphism.common.tsx
@@ -1,0 +1,5 @@
+import { initTRPC } from '@trpc/server';
+
+export const t = initTRPC.create();
+
+export type $RootTypes = (typeof t)['_config']['$types'];
