+++ packages/tanstack-react-query/test/polymorphism.test.tsx
@@ -8,21 +8,24 @@
   The polymorphism types can be used to generate abstract types which routers sharing a common 
   interface are compatible with, and allow you to pass around deep router paths to generic components with ease.
 */
-import { routerToServerAndClientNew } from '../___testHelpers';
+// import { routerToServerAndClientNew } from '../___testHelpers';
+import { getServerAndReactClient } from './__helpers';
 import {
   QueryClient,
   QueryClientProvider,
   useMutation,
+  useQuery,
+  useQueryClient,
 } from '@tanstack/react-query';
 import { render, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import { getUntypedClient } from '@trpc/client';
-import { createTRPCReact } from '@trpc/react-query';
-import type { InferQueryLikeData } from '@trpc/react-query/shared';
 import { konn } from 'konn';
 import type { ReactNode } from 'react';
 import React, { useState } from 'react';
+import { describe, expect, expectTypeOf, test, vi } from 'vitest';
 import { z } from 'zod';
+import type { InferOutput } from '../src';
+import { createTRPCContext } from '../src';
 import { t } from './polymorphism.common';
 /**
  * We define a router factory which can be used many times.
@@ -30,11 +33,7 @@ import { t } from './polymorphism.common';
  * This also exports types generated by RouterLike and UtilsLike to describe
  * interfaces which concrete router instances are compatible with
  */
-import * as Factory from './polymorphism.legacy.factory';
-/**
- * We also define a factory which extends from the basic Factory with an entity sub-type and extra procedure
- */
-import * as SubTypedFactory from './polymorphism.legacy.subtyped-factory';
+import * as Factory from './polymorphism.factory';
 
 /**
  * The tRPC backend is defined here
@@ -48,8 +47,6 @@ function createTRPCApi() {
    */
   const IssueExportsProvider: Factory.FileExportStatusType[] = [];
   const DiscussionExportsProvider: Factory.FileExportStatusType[] = [];
-  const PullRequestExportsProvider: SubTypedFactory.SubTypedFileExportStatusType[] =
-    [];
 
   /**
    * Create an AppRouter instance, with multiple routes using the data export interface
@@ -74,12 +71,6 @@ function createTRPCApi() {
           }),
         ),
       }),
-      pullRequests: t.router({
-        export: SubTypedFactory.createSubTypedExportRoute(
-          t.procedure,
-          PullRequestExportsProvider,
-        ),
-      }),
     }),
   });
 
@@ -88,7 +79,6 @@ function createTRPCApi() {
     appRouter,
     IssueExportsProvider,
     DiscussionExportsProvider,
-    PullRequestExportsProvider,
   };
 }
 
@@ -98,36 +88,14 @@ describe('polymorphism', () => {
    */
   const ctx = konn()
     .beforeEach(() => {
-      const {
-        appRouter,
-        IssueExportsProvider,
-        DiscussionExportsProvider,
-        PullRequestExportsProvider,
-      } = createTRPCApi();
+      const { appRouter, IssueExportsProvider, DiscussionExportsProvider } =
+        createTRPCApi();
+      const opts = getServerAndReactClient(appRouter);
 
-      const opts = routerToServerAndClientNew(appRouter);
-      const trpc = createTRPCReact<typeof appRouter>();
-
-      const queryClient = new QueryClient();
-
-      function App(props: { children: ReactNode }) {
-        return (
-          <trpc.Provider
-            {...{ queryClient, client: getUntypedClient(opts.client) }}
-          >
-            <QueryClientProvider client={queryClient}>
-              {props.children}
-            </QueryClientProvider>
-          </trpc.Provider>
-        );
-      }
       return {
         ...opts,
-        App,
-        trpc,
         IssueExportsProvider,
         DiscussionExportsProvider,
-        PullRequestExportsProvider,
       };
     })
     .afterEach(async (opts) => {
@@ -137,25 +105,25 @@ describe('polymorphism', () => {
 
   describe('simple factory', () => {
     test('can use a simple factory router with an abstract interface', async () => {
-      const { trpc } = ctx;
+      const { useTRPC } = ctx;
 
       /**
        * Can now define page components which re-use functionality from components,
        * and pass the specific backend functionality which is needed need
        */
       function IssuesExportPage() {
-        const utils = trpc.useUtils();
+        const trpc = useTRPC();
+        const client = useQueryClient();
 
         const [currentExport, setCurrentExport] = useState<number | null>(null);
-
         const invalidate = useMutation({
-          mutationFn: () => utils.invalidate(),
+          mutationFn: () => client.invalidateQueries(trpc.github.queryFilter()),
         });
+
         return (
           <>
             <StartExportButton
               route={trpc.github.issues.export}
-              utils={utils.github.issues.export}
               onExportStarted={setCurrentExport}
             />
 
@@ -202,21 +170,22 @@ describe('polymorphism', () => {
     });
 
     test('can use the abstract interface with a factory instance which has been merged with some extra procedures', async () => {
-      const { trpc } = ctx;
+      const { useTRPC } = ctx;
 
       function DiscussionsExportPage() {
-        const utils = trpc.useUtils();
+        const trpc = useTRPC();
+        const client = useQueryClient();
 
         const [currentExport, setCurrentExport] = useState<number | null>(null);
+
         const invalidate = useMutation({
-          mutationFn: () => utils.invalidate(),
+          mutationFn: () => client.invalidateQueries(trpc.github.queryFilter()),
         });
 
         return (
           <>
             <StartExportButton
               route={trpc.github.discussions.export}
-              utils={utils.github.discussions.export}
               onExportStarted={setCurrentExport}
             />
 
@@ -262,86 +231,6 @@ describe('polymorphism', () => {
       });
     });
   });
-
-  describe('sub-typed factory', () => {
-    test('can use a sub-typed factory router with the interfaces from the supertype', async () => {
-      const { trpc } = ctx;
-
-      /**
-       * Can define page components which operate over interfaces generated by a super-typed router,
-       * but also extend types and functionality
-       */
-      function PullRequestsExportPage() {
-        const utils = trpc.useUtils();
-
-        const [currentExport, setCurrentExport] = useState<number | null>(null);
-        const invalidate = useMutation({
-          mutationFn: () => utils.invalidate(),
-        });
-
-        return (
-          <>
-            {/* Some components may still need to be bespoke... */}
-            <SubTypedStartExportButton
-              route={trpc.github.pullRequests.export}
-              utils={utils.github.pullRequests.export}
-              onExportStarted={setCurrentExport}
-            />
-
-            <RefreshExportsListButton
-              mutate={invalidate.mutate}
-              isPending={invalidate.isPending}
-            />
-
-            <RemoveExportButton
-              remove={trpc.github.pullRequests.export.delete}
-              utils={utils.github.pullRequests.export}
-              exportId={currentExport}
-            />
-
-            {/* ... or you can adapt them to support sub-types */}
-            <ExportStatus
-              status={trpc.github.pullRequests.export.status}
-              //                                       ^?
-              currentExport={currentExport}
-              renderAdditionalFields={(data) => {
-                expectTypeOf(data.description).toEqualTypeOf<string>();
-                return `Description: "${data?.description}"`;
-              }}
-            />
-
-            <ExportsList list={trpc.github.pullRequests.export.list} />
-          </>
-        );
-      }
-
-      /**
-       * Test Act & Assertions
-       */
-
-      const $ = render(
-        <ctx.App>
-          <PullRequestsExportPage />
-        </ctx.App>,
-      );
-
-      await userEvent.click($.getByTestId('startExportBtn'));
-
-      await waitFor(() => {
-        expect($.container).toHaveTextContent(
-          'Last Export: `Search for Polymorphism React` (Working)',
-        );
-      });
-
-      await userEvent.click($.getByTestId('refreshBtn'));
-
-      await waitFor(() => {
-        expect($.container).toHaveTextContent(
-          'Last Export: `Search for Polymorphism React` (Ready!)',
-        );
-      });
-    });
-  });
 });
 
 /**
@@ -352,84 +241,27 @@ describe('polymorphism', () => {
 
 function StartExportButton(props: {
   route: Factory.ExportRouteLike;
-  utils: Factory.ExportUtilsLike;
   onExportStarted: (id: number) => void;
 }) {
-  const exportStarter = props.route.start.useMutation({
-    async onSuccess(data) {
-      props.onExportStarted(data.id);
+  const client = useQueryClient();
 
-      await props.utils.invalidate();
-    },
-  });
+  const exportStarter = useMutation(
+    props.route.start.mutationOptions({
+      async onSuccess(data) {
+        props.onExportStarted(data.id);
+
+        await client.invalidateQueries(props.route.queryFilter());
+      },
+    }),
+  );
 
   return (
     <button
       data-testid="startExportBtn"
       onClick={() => {
-        exportStarter.mutateAsync({
+        exportStarter.mutate({
           filter: 'polymorphism react',
           name: 'Search for Polymorphism React',
-        });
-      }}
-    >
-      Start Export
-    </button>
-  );
-}
-
-function RemoveExportButton(props: {
-  exportId: number | null;
-  remove: SubTypedFactory.ExportSubTypedRouteLike['delete'];
-  utils: SubTypedFactory.ExportSubTypesUtilsLike;
-}) {
-  const exportDeleter = props.remove.useMutation({
-    async onSuccess() {
-      await props.utils.invalidate();
-    },
-  });
-
-  const id = props.exportId;
-  if (!id) {
-    return null;
-  }
-
-  return (
-    <button
-      data-testid="removeExportBtn"
-      onClick={() => {
-        exportDeleter.mutateAsync({
-          id,
-        });
-      }}
-    >
-      Remove Export
-    </button>
-  );
-}
-
-type SubTypedStartExportButtonProps = {
-  route: SubTypedFactory.ExportSubTypedRouteLike;
-  utils: SubTypedFactory.ExportSubTypesUtilsLike;
-  onExportStarted: (id: number) => void;
-};
-function SubTypedStartExportButton(props: SubTypedStartExportButtonProps) {
-  const exportStarter = props.route.start.useMutation({
-    onSuccess(data) {
-      props.onExportStarted(data.id);
-
-      props.utils.invalidate();
-    },
-  });
-
-  return (
-    <button
-      data-testid="startExportBtn"
-      onClick={() => {
-        exportStarter.mutateAsync({
-          filter: 'polymorphism react',
-          name: 'Search for Polymorphism React',
-          description: 'This field is unique to the sub-typed router',
         });
       }}
     >
@@ -457,12 +289,14 @@ function ExportStatus<
   TStatus extends Factory.ExportRouteLike['status'],
 >(props: {
   status: TStatus;
-  renderAdditionalFields?: (data: InferQueryLikeData<TStatus>) => ReactNode;
+  renderAdditionalFields?: (data: InferOutput<TStatus>) => ReactNode;
   currentExport: number | null;
 }) {
-  const exportStatus = props.status.useQuery(
-    { id: props.currentExport ?? -1 },
-    { enabled: props.currentExport !== null },
+  const exportStatus = useQuery(
+    props.status.queryOptions(
+      { id: props.currentExport ?? -1 },
+      { enabled: props.currentExport !== null },
+    ),
   );
 
   if (!exportStatus.data) {
@@ -473,13 +307,15 @@ function ExportStatus<
     <p>
       Last Export: `{exportStatus.data?.name}` (
       {exportStatus.data.downloadUri ? 'Ready!' : 'Working'})
-      {props.renderAdditionalFields?.(exportStatus.data as unknown as any)}
+      {props.renderAdditionalFields?.(exportStatus.data)}
     </p>
   );
 }
 
-function ExportsList(props: { list: Factory.ExportRouteLike['list'] }) {
-  const exportsList = props.list.useQuery();
+function ExportsList(props: {
+  list: Omit<Factory.ExportRouteLike['list'], 'nothing'>;
+}) {
+  const exportsList = useQuery(props.list.queryOptions());
 
   return (
     <>
